% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_func.R
\name{addCaveats}
\alias{addCaveats}
\title{Add caveats function}
\usage{
addCaveats(new_caveat)
}
\arguments{
\item{new_caveat}{The added caveat to put into the caveats folder}
}
\value{
None, assigns the caveat to the environment
}
\description{
Add caveats function
}
