% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_func.R
\name{data_type_format}
\alias{data_type_format}
\title{One Stop Data Type Formatting}
\usage{
data_type_format(
  dfin,
  dateform = "\%Y-\%m-\%d",
  datereg = "date",
  numreg = "value",
  flagreg = "flag"
)
}
\arguments{
\item{dfin}{The input dataframe}

\item{dateform}{The format of any dates in the dataframe as a string}

\item{datereg}{The regular expression to determine if a column name justifies treating it as a date}

\item{numreg}{The regular expression to determine if a column name justifies treating it as numeric}

\item{flagreg}{The regular expression to determine if a column name justifies treating it as a factor}
}
\value{
The dataframe with the appropriate data type transformations applied to the columns
}
\description{
A function to perform the routine data type transformations at the beggining of a data flow in one easy go.
Most data is read into R as character data, but most useful data we work with is not character data,
this package allows you to direct R to transform all the columns suiting a regular expression you define to
be transformed as either date, numeric or factor data.
}
