% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_func.R
\name{left_join_monitor}
\alias{left_join_monitor}
\title{Monitored Left Join}
\usage{
left_join_monitor(x, y, by, ...)
}
\arguments{
\item{x}{The LHS object}

\item{y}{The RHS object}

\item{by}{the joining column(s) between x & y}

\item{...}{Additional argiments to be passed to dplyr::nest_join}
}
\value{

}
\description{
Monitored Left Join
}
