% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{osb_DotPlot}
\alias{osb_DotPlot}
\title{generate dot plot}
\usage{
osb_DotPlot(dataset, changes_in, for_the, across, formatter = function(x) x)
}
\arguments{
\item{formatter}{}
}
\value{

}
\description{
generate dot plot
}
