% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{osb_DriftPlot}
\alias{osb_DriftPlot}
\title{generate drift plot}
\usage{
osb_DriftPlot(dataset, changes_in, for_the, across, formatter = function(x) x)
}
\arguments{
\item{formatter}{}
}
\value{

}
\description{
generate drift plot
}
