% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{osb_SlopePlot}
\alias{osb_SlopePlot}
\title{generate slope plot}
\usage{
osb_SlopePlot(dataset, changes_in, for_the, across)
}
\arguments{
\item{dataset}{The source dataset}

\item{changes_in}{The variable to show changes in}

\item{for_the}{The variable to split those changes by}

\item{across}{The two timeframes for the changes to occur}
}
\value{

}
\description{
generate slope plot
}
