% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_func.R
\name{printAndIncrementStep}
\alias{printAndIncrementStep}
\title{Step incrementer and assigner}
\usage{
printAndIncrementStep(val)
}
\arguments{
\item{val}{the object to step}
}
\value{
the value of that object, the object has been incremented in the background
}
\description{
Step incrementer and assigner
}
