% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_func.R
\name{showIndistinct}
\alias{showIndistinct}
\title{show the indistinct values}
\usage{
showIndistinct(df, ...)
}
\arguments{
\item{df}{Source Dataframe}

\item{...}{The variable, or variables which you want to view all indistinct combinations of.}
}
\value{

}
\description{
show the indistinct values
}
