% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_caster.R
\name{util_IndexCaster}
\alias{util_IndexCaster}
\title{Index Caster Function}
\usage{
util_IndexCaster(.index)
}
\arguments{
\item{.index}{The supplied index value to be converted}
}
\description{
A function to cast an inde as an measure of its deviation from a
standard, circumventing the issue of fractional changes in negative indicies.
}
\examples{
     util_IndexCaster(50)

     util_IndexCaster(200)

}
