% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_func.R
\name{utilRange}
\alias{utilRange}
\title{Get the range of a numeric vector}
\usage{
utilRange(a_vect)
}
\arguments{
\item{a_vect}{a numeric vector}
}
\value{

}
\description{
Get the range of a numeric vector
}
