% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namespace_utils.R
\name{get_used_namespace}
\alias{get_used_namespace}
\title{Get the currently used namespace}
\usage{
get_used_namespace(.function_name)
}
\arguments{
\item{.function_name}{}
}
\value{

}
\description{
Function to get the used namespace and prepend it to the
  function name
}
