% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_func.R
\name{is.distinct}
\alias{is.distinct}
\title{Distinctness checker}
\usage{
is.distinct(df, on)
}
\arguments{
\item{df}{Source dataframe}

\item{on}{The NSE variable which you want to check the distinctiveness of}
}
\value{

}
\description{
Distinctness checker
}
