% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{osb_DriftPlot}
\alias{osb_DriftPlot}
\title{generate drift plot}
\usage{
osb_DriftPlot(dataset, changes_in, for_the, across, formatter = function(x) x)
}
\arguments{
\item{dataset}{The source dataset}

\item{changes_in}{The variable to show changes in}

\item{for_the}{The variable to split those changes by}

\item{across}{The two timeframes for the changes to occur}

\item{formatter}{A formatter function to control the labels}
}
\value{

}
\description{
generate drift plot
}
