% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_func.R
\name{printStepTitle}
\alias{printStepTitle}
\title{Step title printer}
\usage{
printStepTitle(string)
}
\arguments{
\item{string}{The title of the step}
}
\value{
NULL The function instead prints the step title along with a
  timestamp
}
\description{
This function is a bit dodge, relies on a variable called
  analysis step in the environment, quite a niche application, but relied on
  in some existing code so im going to leave it in.
}
